% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{mk_df}
\alias{mk_df}
\title{Make Data Frame}
\usage{
mk_df(attributes, object)
}
\arguments{
\item{attributes}{column vector containing the names of desired metadata}

\item{object}{A \code{RangedSummarizedExperiment} object created by running regularized log transformation \code{DESeq2::rlog()} or
variance stabilizing transformation \code{DESeq2::vsd()} on a \code{DESeqDataSet} object.}
}
\value{
Returns a data frame that will be used as \code{df} argument in \code{mk_heatmap}
}
\description{
Creates the legend for the heatmap
}
\examples{
## Example 1

# Creates a data frame using all metadata given from your design table in your 'object'
> df <- mk_df(attributes = c("replicate", "sample", "allele"), object = rld)

## Example 2

# Creates a data frame using only two attributes give form your design table in your 'object'
> df <- mk_df(attributes = c("replicate", "sample"), object = rld)

}
