% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{mk_MAplot}
\alias{mk_MAplot}
\title{Make Mean-Average Plot}
\usage{
mk_MAplot(
  format,
  res,
  pCutoff,
  geneList,
  ntop = 20,
  title,
  subtitle,
  unit,
  ylim
)
}
\arguments{
\item{format}{Specifies how to subset the \code{res}. \code{"list"} will use desired genes symbols while
\code{"padj"} will use the lowest \code{'padj'} values .}

\item{res}{Contrast data produced by \code{DESeq2::res()} or \code{DESeq2::lfcShrink()}}

\item{geneList}{Input, desired list of genes}

\item{ntop}{Limits the number of genes included in \code{format = "padj"}.}

\item{title}{Input title for graph}

\item{subtitle}{Input subtitle for graph}

\item{unit}{Labels graph to denote the comparison used for \code{res}}

\item{ylim}{Defines the Y-Axis boundaries}

\item{pCutOff}{P-adj threshhold. Any values above \code{pCutOff} will be marked in grey, otherwise red on the graph}
}
\description{
Modified plotMA from DESeq2 package. Circles and labels either genes with the lowest p-adj values or specific genes
}
\examples{
## Example 1: Using padj

# MA-plot with the lowest p-adj genes circled (defaults to 20 genes)
> mk_MAplot(format = "padj", res = res pCutOff = 0.05)

# Specifying lowest 20 p-adj genes 
> mk_mat(format = "padj", res = res, pCutoff = 0.05, ntop = 10)

# Setting the boundaries of the y-axis to [-5,5]
# Please make sure the ylim[1] is the lower bound while ylim[2] is the upper bound
> mk_mat(format = "padj", res = res, pCutoff = 0.05, ylim = c(-5,5))

## Example 2: Using list

# Creates a matrix by subsetting the 'rld' with desired list of genes
> selectLab <- c("TSPAN6", "TNMD", "DPM1", "SCYL3", "C1orf112", "FGR")
> mat <- mk_mat(format = "list", res = res, geneList = selectLab)

## Example 3: Exporting graph to pdf
> pdf("MAPlot.pdf")
> mk_MAplot(format = "padj", res = res pCutOff = 0.05)
> dev.off()
}
