% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{mk_mat}
\alias{mk_mat}
\title{Make Matrix}
\usage{
mk_mat(format, object, geneList, res, pCutoff, nTop = 100)
}
\arguments{
\item{format}{Specifies how to subset the matrix. \code{"list"} will produce a matrix with only desired genes symbols. \code{"padj"} will
subset \code{object} based on the metadata \code{'padj'} provided by \code{res}.}

\item{object}{A \code{RangedSummarizedExperiment} object created by running regularized log transformation \code{DESeq2::rlog()} or
variance stabilizing transformation \code{DESeq2::vsd()} on a \code{DESeqDataSet} object.}

\item{geneList}{Desired list of gene symbols}

\item{res}{\code{DESeqResults} object \code{DESeq2::results()} containing the necessary metadata to subset \code{object}}

\item{nTop}{Limits the number of genes that are included in the matrix in \code{format = "padj"}, by default this value is set to 100.}

\item{pCutOff}{P-adj threshhold. Any genes above \code{pCutOff} will not be included in the matrix}
}
\value{
Returns a matrix that replaces ensembl IDs with gene symbols. Additionally, the matrix that highlights the the upregulation
and downregulation of gene expression across samples, the values within each entry are subtracted from the entry's mean values. That way,
upregulation are indicated by positive values while downregulation are indicated by negative values.
}
\description{
Creates a matrix that highlights the upregulation and downregulation of either sigificant genes or specific
genes across all samples.
}
\examples{
## Example 1: Using padj

# Creates a matrix by subsetting the 'rld' the with p-adj threshhold of 0.05 (defaults to 100 rows)
> mat <- mk_mat(format = "padj", object = rld, res = res, pCutoff = 0.05)

# Creates a matrix by subsetting the 'rld' with 20 lowest p-adj genes, granted the values are under 0.05
> mat <- mk_mat(format = "padj", object = rld, res = res, pCutoff = 0.05, nTop = 20))

## Example 2: Using list

# Creates a matrix by subsetting the 'rld' with desired list of genes
> selectLab <- c("TSPAN6", "TNMD", "DPM1", "SCYL3", "C1orf112", "FGR")
> mat <- mk_mat(format = "list", object = rld, geneList = selectLab)
}
