% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{mk_pca}
\alias{mk_pca}
\title{Make Principle Component Analysis}
\usage{
mk_pca(object, color, label, shape, ntop = 500)
}
\arguments{
\item{object}{A \code{RangedSummarizedExperiment} object created by running regularized log transformation \code{DESeq2::rlog()} or
variance stabilizing transformation \code{DESeq2::vsd()} on a \code{DESeqDataSet} object.}

\item{color}{Desired intgroup/attribute from \code{RangedSummarizedExperiment::colData(obj)} you want denoted with color}

\item{label}{Desired intgroup/attribute from \code{RangedSummarizedExperiment::colData(obj)} you want denoted with labels}

\item{shape}{Desired intgroup/attribute from \code{RangedSummarizedExperiment::colData(obj)} you want denoted with shapes}
}
\value{
Returns PCA either in your plots window, alternatively can be directed into a file with \code{pdf(file = "file_name")} and \code{dev.off()}
}
\description{
Modified plotPCA from DESeq2 package that allows you to denote attributes/interest groups
with color, labels, or shapes.
}
\examples{
> colnames(SummarizedExperiment::colData(rld)) 
[1] "replicate"  "sample"     "allele"     "sizeFactor"

## Example 1 Label Sample Names
> mk_pca(rld)

## Example 2 Highlight intgroup "replicate" by color
> mk_pca(rld, color = "replicate")

## Example 3 Label samples by intgroup "replicate"
> mk_pca(rld, label = "replicate")

## Example 4  Highlight all interest group by color, label, and shape
> mk_pca(rld, color = "replicate", label = "sample", shape = "allele")

## Example 5 Exporting graph into a pdf 
> pdf("PCA.pdf")
> mk_pca(rld, color = "replicate", label = "sample", shape = "allele")
> dev.off()
}
